Promoted second set of files to main set
