--- youtube-playback-2.js
@@ -1,2 +1,0 @@
-// Dynamic import to set up environment as a module and add import metadata
-import('https://calebanderson.github.io/youtube-playback-2/core.js');
