--- youtube-playback-2/core.js
@@ -1,22 +1,0 @@
-import './remove_shim.js';
-import CustomVideo from './custom_video.js';
-
-// Wanted to try out promises... obviously not the most useful implementation...
-function repeatUntilPresent(func, delay){
-  const checkForPresence = function(resolve){
-    const result = func();
-    if(result){
-      resolve(result);
-    } else {
-      setTimeout(checkForPresence.bind(this, resolve), delay);
-    }
-  }
-  return new Promise(checkForPresence);
-}
-
-Promise.all([
-  repeatUntilPresent(()=>document.querySelector(CustomVideo.videoElementSelector), 250),
-  repeatUntilPresent(()=>document.querySelector(CustomVideo.badgeContainerSelector), 250)
-]).then(([video, container])=>{
-  window.customVideo = new CustomVideo(video);
-});
