--- youtube-playback-2/custom_toggle_button.js
@@ -1,13 +1,0 @@
-import CustomBadge from './custom_badge.js';
-
-class CustomToggleButton extends CustomBadge {
-  connectedCallback(){
-    super.connectedCallback();
-
-    this.classList.add('toggle-button');
-    this.classList.add('clickable');
-  }
-}
-customElements.define('custom-toggle-button', CustomToggleButton);
-
-export default CustomToggleButton;
